fix(dispatch): keep thread preview helpers browser safe
