+++ .changeset/tidy-buses-serve.md
@@ -1,0 +1,5 @@
+---
+"@agent-native/dispatch": patch
+---
+
+Keep Dispatch thread preview route metadata browser-safe by moving pure preview helpers out of server modules.
