+++ packages/dispatch/package.json
@@ -22,8 +22,10 @@
     "./routes": "./dist/routes/index.js",
     "./routes/pages/*": "./dist/routes/pages/*.js",
     "./server": "./dist/server/index.js",
+    "./server/lib/thread-link-preview": "./dist/server/lib/thread-link-preview.js",
     "./actions": "./dist/actions/index.js",
     "./db": "./dist/db/index.js",
+    "./lib/thread-link-preview": "./dist/lib/thread-link-preview.js",
     "./components": "./dist/components/index.js",
     "./components/ui/*": "./dist/components/ui/*.js",
     "./styles/dispatch.css": "./src/styles/dispatch.css"
