+++ packages/dispatch/src/routes/pages/overview.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useMemo, useState } from "react";
-import { Link, useNavigate, type LoaderFunctionArgs } from "react-router";
+import { Link, useNavigate } from "react-router";
 import {
   PromptComposer,
   useActionQuery,
@@ -34,10 +34,6 @@ import {
   TooltipTrigger,
 } from "@/components/ui/tooltip";
 import { submitOverviewPrompt } from "@/lib/overview-chat";
-import {
-  buildThreadLinkPreviewMeta,
-  loadThreadLinkPreview,
-} from "@/server/lib/thread-link-preview";
 import type { WorkspaceAppSummary } from "@/lib/workspace-apps";
 
 interface IntegrationStatus {
@@ -475,17 +471,8 @@ function StepRow({ step }: { step: ChecklistStep }) {
   );
 }
 
-export async function loader({ request }: LoaderFunctionArgs) {
-  const threadId = new URL(request.url).searchParams.get("thread");
-  return {
-    threadPreview: await loadThreadLinkPreview(threadId),
-  };
-}
-
-export function meta({ data }: { data?: Awaited<ReturnType<typeof loader>> }) {
-  return data?.threadPreview
-    ? buildThreadLinkPreviewMeta(data.threadPreview)
-    : [{ title: "Overview — Dispatch" }];
+export function meta() {
+  return [{ title: "Overview — Dispatch" }];
 }
 
 export default function OverviewRoute() {
